fix(ui): stop an infinite loop opening a rule's detail page near port 65535
ruleDetailView built the split-point option list with
"for p := rule.ListenPort.Lo + 1; p <= rule.ListenPort.Hi; p++" over
uint16 p. When Hi is 65535, incrementing p past it wraps to 0, so
p <= Hi holds forever and the request never returns. A rule with
listen_port such as 65534-65535 reproduces it. Loop over int instead
and convert back to uint16 per port.

Checked the rest of the tree for the same uint16 loop-increment
pattern (grep for "Lo + 1"/"Hi+1" and uint16 for-loops): the userspace
dataplane and relay manager already loop with int, and the two Hi+1
adjacency comparisons in proto/splitmerge.go and webui_splitmerge.go
wrap to 0 only when Hi is 65535, which correctly compares unequal to
any valid Lo (>=1) rather than looping, so no further fix was needed
there.

+++ internal/vpsd/admin/webui_rule.go
@@ -182,8 +182,11 @@ func (s *Server) ruleDetailView(rule proto.Rule, locale string) ruleDetailData {
 	d.CanSplit = rule.ListenPort.IsRange()
 	if d.CanSplit {
 		d.ListenLo, d.ListenHi = rule.ListenPort.Lo, rule.ListenPort.Hi
-		for p := rule.ListenPort.Lo + 1; p <= rule.ListenPort.Hi; p++ {
-			d.SplitPorts = append(d.SplitPorts, p)
+		// int で回す。p を uint16 のまま Hi(最大 65535)まで回すと、p が 65535 に達した
+		// 直後の p++ が 0 に折り返り、p <= Hi が恒に真になって無限ループになる
+		// (listen_port=65534-65535 のようなルールの詳細ページを開くと再現する)。
+		for p := int(rule.ListenPort.Lo) + 1; p <= int(rule.ListenPort.Hi); p++ {
+			d.SplitPorts = append(d.SplitPorts, uint16(p))
 		}
 		if host, portStr, err := net.SplitHostPort(rule.Target); err == nil {
 			if port, err := strconv.ParseUint(portStr, 10, 16); err == nil {
